Assignment 2
In the assignment, you will be required to explore, segment, and cluster
the neighborhoods in the city of Toronto. However, unlike New York, the
neighborhood data is not readily available on the internet, so you need
to learn to be agile and refine the skill to learn new libraries and
tools quickly depending on the project.

For the Toronto neighborhood data, a Wiki page exists that has all the
information we need to explore and cluster the neighborhoods in Toronto.
You will be required to scrape the Wikipedia page and wrangle the data,
clean it, and then read it into a pandas dataframe so that it is in a
structured format like the New York dataset.

**Type** feat: A new feature for the user.